Maintain the release notes on the homepage
By moving the release notes to the homepage we make the process of
publishing them easier. Edits and modifications can be made in the
online editor and after being submitted they are immediately visible
on the homepage. This also means that release notes for upcoming
releases are made visible earlier and users can see what new features
are coming.

We also don't need to maintain a build system for the release notes
in the core Gerrit repository, and maintainers don't need to manually
generate and upload them to the Google Storage bucket for each release.

In this change the release notes for the 2.12.x and 2.13.x releases
are migrated. A link to the previous release notes site is given for
older releases.

Change-Id: I4ee59f1a9d967dba86a5f7c555cfe5a246432bd9