RFC: Maintain the release notes on the homepage
There are several benefits of moving the release notes to the homepage:

- Edits and modifications can be made in the online editor and after
  review and submit they are immediately visible on the homepage.

- We don't need to maintain a build system for the release notes in
  the core Gerrit repository.

- Maintainers don't need to manually generate the release notes and
  upload them to the Google Storage bucket for each release.

- The markdown formatting and presentation is nicer than asciidoctor.

- All of the above lowers the bar for other contributors to make
  changes in the release notes.

In this RFC only the 2.13.1 release notes are ported. If this idea
is accepted, I will do the remaining ones too.

Eventually we could probably also replace the existing external
downloads page with a page in the homepage.

Change-Id: I4ee59f1a9d967dba86a5f7c555cfe5a246432bd9