Update patch set 1
Patch Set 1:

> Even though elasticsearch has not been fully implemented. It still
 > works in some ways (haven't tested, but i belive wikimedia will use
 > it when they upgrade to 2.14)
 > 

I don't recommend they use it until the remaining parts are implemented.   The major missing part is online reindex support, so if that's missing the entire reindex has to be done offline.

 > It's best to note that this could be a beta product or that not all
 > of the implementation has been completed.

Possibly, but if that's the case I would prefer to put that back in the release note at a later date.  Right now I don't want people reading it and assuming it's going to be in.

Patch-set: 1
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>