Update patch set 38
Patch Set 38:

> It's difficult to update the screenshots, as PolyGerrit isn't yet (I don't think?) the default.

Sure, but:

The screenshots are using a version of the UI that doesn't exist any more (i.e. they do not match what you get by clicking "Switch to Old UI" at the bottom of this screen). That makes them confusing to use. Either the current "old UI" or PolyGerrit would be less confusing. There are pros and cons to each: the "old UI" is not likely to change soon, making screenshots of it easier to maintain than screenshots of PolyGerrit; on the other hand, PolyGerrit is more intuitive and will become the default soon.

Patch-set: 38
CC: Gerrit User 1011842 <1011842@173816e5-2b9a-37c3-8a2e-48639d4f1153>