Update patch set 39
Patch Set 39:

> Patch Set 39:
> 
> (1 comment)
> 
> Is the intention to move Gerrit's documentation out of the core gerrit repository into the homepage?  If so, is it going to be all documentation, or just high level parts?
> 
> What is the plan for maintaining documentation of previous versions?

Excellent question. The short answer is no, not at this time. However, I think users would benefit from having a set of docs that are not part of the product-documentation. Having these guides allows us to better hone the on-boarding experience for new Gerrit users--some of whom may not have an installation to work with yet. Also, I'd like to add Google Analytics to topics so we can see better gauge the impact of our docs. That's difficult to do with docs that are in the core gerrit repo, because we don't know where those docs are being served.

We may eventually decide that it's best to have high level docs in one area, and leave task-focused docs in the core Gerrit repo, as you suggest. However, I don't think we know enough yet to make that kind of decision.

Regarding previous versions of these guides: I think the best thing to do (for now) is either tag what versions the guides support. If it isn't too much overhead, we could store previous guide versions, just as we do for previous versions of the core gerrit docs.

Patch-set: 39
Reviewer: Gerrit User 1041213 <1041213@173816e5-2b9a-37c3-8a2e-48639d4f1153>