Add Shawn Pearce Memorial Fund to our homepage
Shawn is the project founder, and we are all grateful for his
dedication, passion, help, and inspiration for fuelling this
fantastic OpenSource project: Gerrit Code Review.

The minimum we can do to say "Thank you" is to help to support
his family in this painful moment. This is what he wanted from us.

Change-Id: Ie343e4ed9c9804ca76490bcb63bc1108850dd7cb