Add stub page for 2.11 release notes
When we migrated the release notes to the new project homepage, only
the release notes from 2.12 and later were migrated.

Now we're about to do a new release, 2.11.11, and will need a place to
put the release notes.

Add a basic stub page for 2.11, mentioning only the release highlights
and a link to the full release notes on the old site.

The release notes for 2.11.11 will be added in a separate section on
this new page later.

Change-Id: I2950efc7bdd8a74c0e96d5d0253a9556d6b0ecc3