Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This is what I want to see:
> 
> * Make a change on the "jekyll" branch in the homepage repo (which I just created, initially pointing at the same place as md-pages).
> * Edit existing .md files in place, so we can see exactly what's changing when we switch to jekyll.
> * Add new files to set up the jekyll build (Gemfile, Dockerfile, etc.)
> * No jekyll-source directory, everything in the top level. There should be nothing left in this repo that is *not* the jekyll version of the site.
> * This implies that we need to convert the remaining site content (about.md, docs/*, etc.) to jekyll.

Ok. I think I see where you are going with this.

All of the content for the site will be somewhere in the /pages directory. I don't think the subdirectories matter when the site is generated, so what if I created two:

/pages/website - for all website content
/pages/docs/ - for all documentation

The root of the repo will still require all the source files and such that Jekyll needs. That's ok, right?

Patch-set: 1
Reviewer: David Shevitz <1041213@173816e5-2b9a-37c3-8a2e-48639d4f1153>