News index: Fix display of post summary
The template is using 'page.summary' which does not exist, and thus
ends up using an extract of the post's content, rather than the
defined summary. Change it to 'post.summary'.

Also, don't truncate the summary at 160 characters; this can cause
the summary to be cut off unexpectedly. Instead, let's rely on the
reviewers to check that news posts have a summary entry with a
reasonable length.

Change-Id: I970ddffba27132c5a48602f49618b2a5d5a1ca5f