Fix anchors to maintenance release notes
Since the migration to Jekyll, the anchors for the release notes
look like:

  #21412

instead of how they were before:

  #2.14.12

This breaks all the external links (for example from the release
announcement emails), and all the internal links within the release
notes themselves.

Add explicit anchors on the sections, so that now we have both the
ones automatically generated by Jekyll (#21412) and the ones that
were there before (#2.14.12).

Bug: Issue 9675
Change-Id: Ifcca1c30e5d54146cb4390e3a2992658948f77d4