Document which Gerrit versions are actively maintained
The Gerrit open source community only supports the last 2 versions
actively but so far this wasn't documented. Add this information to the
homepage and make clear that all support happens on a best effort basis
as Gerrit is an open source project and we do not offer support
contracts.

Also mention that the repo-discuss mailing list is managed and new posts
need to be approved before they appear on the mailing list. This may
help to avoid duplicate posts from new participants.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Idb7a3c39da0a96c605fd0c4219cc195e7b64ab8a