Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> >  In that case, they can just upgrade the war and restart Gerrit.
> 
> Yes, teh online reindexing is enough, and that's exactly why we decided to bump the index schema version (first time - and as an exception from the rule - on stable branch): to have the best possible upgrade experience for Gerrit users.

It technically works, however, the project reindex takes (at least for us) around 30 mins and, when it is active, is really heavy for the machine. It would be best to run it offline IMHO. The reindex of Accounts and Groups is pretty fast and thus makes sense to run online.

Let me do an experiment to see how long and how slow would it be when upgrading the projects index online.

Patch-set: 1
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>