Update patch set 2
Patch Set 2: Code-Review-1

This doesn't need to be in the release notes.  Release notes are for fixes and features that affect the end-user or have impact on Gerrit at runtime, while the minimum bazel version is only relevant for developers.

If we mention the bazel version anywhere, it should be in dev-bazel.txt in the core gerrit project.

Patch-set: 2
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1