Update patch set 3
Patch Set 3:

> Patch Set 2: Verified+1 Code-Review+2
> 
> The only missing section is the Native packaging one, which I suppose is normal at this early development stage.

The native packaging one was included in 3.0 for the first time due to changes in the versions used.  I don't expect that we will need to make such changes again, but if we do we can add that section.

Patch-set: 3
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>