Add a page listing the community managers
List the community managers for the 2019 term.

Same as the lists of maintainers and ESC members, it is not linked
from any other pages yet. The intention is to link to it from the
community governance documentation in the core Gerrit project, and/or
from a new (yet to be created) governance page on the homepage.

Change-Id: I0fe36b8b25cae6637657b204e2036e80debf9592