Add introductory paragraphs for maintainers and ESC
Add an introductory paragraph, taken from the project governance
documentation in the core Gerrit project, and add links to the
full documentation that is hosted on gerrit-review.

In later changes we may consider moving the entire governance
documentation from core Gerrit to the homepage, but this is beyond
the scope of the current change.

Change-Id: I9617f4622066abc3661d0d8a20cf8ceff108ed6c