Add maintainer affiliations
Add a column on the maintainer list showing the current and
former affiliations of the maintainers.

This makes the list consistent with the list of ESC members, and
the list of community managers which will be added in a follow-up
change.

It also makes it clear to the community what are the affiliations
of the maintainers.

Change-Id: Ifb7e4ae9b9ff3abeb63f509286e080dc34b3facc