Add initial draft post for project news in June 2019
Add the initial skeleton of the post.

The intention is to follow a process similar to how the Git Rev News
is written, i.e. anyone from the community can contribute newsworthy
items throughout the next few weeks, and then a maintainer will tidy
it up and publish it on June 28th.

Change-Id: I91ea31b0cd9a27b146ca9cadec6760c3bb92f836