Fix tags and keywords on posts
The tags are appearing with double-commas on the index.

Update the documentation to say that tags and keywords on blog posts
should be written in the markdown as space-delimited lists, rather
than comma-delimited.

Fix the existing posts accordingly.

There are still some glitches with tag handling, which are not fixed
in this commit:

- The tag archive pages are not being generated, so clicking on any
  of the tag links results in a "not found" page.

- All the tags are rendered on the index, but only the first tag is
  rendered in the actual post.

Change-Id: Ifaa8f2b020e1e8beaac4226658b8fd3435afc0de