Add design doc section to homepage
As part of being a more transparent community, we wanted to create a
place to publicly review design docs of larger features before
implementing them.

This gives community members a chance to participate in the design
process and to raise concerns if there are any.

This commit adds an index page as a place for future design docs to
be added.

Change-Id: Ia3eb7fdc677a34856c674dc2808fc3ebc0198eeb