News: Add post to announce 2019 Community Managers
As this was not formally done yet. Introduce the Community backlog in
Monorail and confidential group for community members feedback. Mention
these being added to the Gerrit documentation "now" (once this posted).

Introduce first community survey for members to start answering at will.

Mean to share this news post and survey with the community by posting
its link to the mailing list once published.

Bug: Issue 10978, Issue 10979
Change-Id: Ifa00e4055b1e8037c3a38195c3ebd9340aa81c24