News: Add post to announce 2019 Community Managers
As this was not formally done yet. Introduce the Community backlog in
Monorail and confidential group for community members feedback. Mention
these being added to the Gerrit documentation "now" (once this posted).

Introduce the first community survey for members to start answering at
will. Anonymously accept answers between now and July 2019 included,
prior to the 2019 community retrospective in August. This survey is
based on questions from [1] below, adapted from generic teams to the
Gerrit community context. That article was deemed authoritative here.

Mean to share this news post and survey with the community by posting
its link to the mailing list once published.

[1] https://medium.com/the-liberators/agile-teams-dont-use-happiness-metrics-measure-team-morale-3050b339d8af

Feature: Issue 10978, Issue 10979
Change-Id: Ifa00e4055b1e8037c3a38195c3ebd9340aa81c24