Update patch set 5
Patch Set 5:

(2 comments)

For completeness of the picture and because I see a lot of discussion around loose vs. packed refs that made it sounds like this is an issue of this design in particular: ExternalIDs and Accounts are currently updated in a BatchRefUpdate, so they suffer from the limitations that Martin outlined around packed-refs.

We might have a shot at fixing that now.

Patch-set: 5
Reviewer: Gerrit User 1026112 <1026112@173816e5-2b9a-37c3-8a2e-48639d4f1153>