Update patch set 5
Patch Set 5:

(1 comment)

> Patch Set 5:
> 
> (2 comments)
> 
> For completeness of the picture and because I see a lot of discussion around loose vs. packed refs that made it sounds like this is an issue of this design in particular: ExternalIDs and Accounts are currently updated in a BatchRefUpdate, so they suffer from the limitations that Martin outlined around packed-refs.
> 
> We might have a shot at fixing that now.

It seems to me that the current storage format does not actually dictate this, does it? As long as the ExternalIds ref is updated first, it should avoid any "primary key" issues and the user account ref could be safely updated independently. Perhaps this really is just a miss-use of the BatchRefUpdate? Is there a valid reason to need it? Perhaps fixing this would fix the contention issues that Luca is currently seeing?

Patch-set: 5
Reviewer: Gerrit User 1003883 <1003883@173816e5-2b9a-37c3-8a2e-48639d4f1153>