Update patch set 1
Patch Set 1: Code-Review-2

We discussed this design in yesterday’s ESC meeting.

Adding the tree API to core Gerrit has a cost which isn't balanced sufficiently with benefits for Gerrit in our opinion.

If we added this new API, we would need to maintain it in the future and also make sure that other features interact nicely with it. We tried to come up with other use cases which could use this API (as part of Gerrit or other tools of the Gerrit community) but didn't find any. As your comments sound as if GPG-signed commits of UI actions could never be implemented directly in Gerrit (not even as a frontend part/plugin), there would never be a use of this API in Gerrit. This would also make it difficult in the future for us if we ever wanted to change or remove this API as there wouldn't be a direct contact in Gerrit's community as a main stakeholder of this API.

As mentioned in some comments, there are alternatives to this new API being part of core Gerrit (e.g. using plain Git in the way Jonathan suggested; asking host owners to install the Gitiles plugin). Those also offer some benefits. For instance, plain Git might be much more efficient and less resource-intensive (if used with the right set of options). The Gitiles plugin is also available for older versions of Gerrit and hence users of an older Gerrit host could also benefit from your solution.

Of course, there might always be users who prefer one over the other options as their specific Gerrit setup only supports one of them (e.g. no Gitiles plugin; old Git version; old Gerrit version). If there was a large Gerrit user base who was using your Chrome extension, we could optimize for one of them. At the moment, we don’t think that we are in that situation. If this changes and if there was also a permanent stakeholder for this in Gerrit’s community, we would be happy to re-evaluate the addition of this API.

Patch-set: 1
Reviewer: Gerrit User 1030912 <1030912@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-2