Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> > Patch Set 1:
> > 
> > > 
> > > By the way, I followed the discussion on the chat, and I believe this will be an unannounced release and we'll create another 2.16.11.1, right?
> > 
> > Yes, which I guess means this change will need to be abandoned or at least adapted to say that 2.16.1 is broken.  We've done the latter in a couple of cases before.
> 
> Ah, in that case, if 2.16.11 would not be available at all, then this change should not be merged as-is, yeah.

Actually the current situation is a bit different to previous cases.  Previously we found issues after the release had been published and announced, and then a follow up release was made.

In this case 2.16.11 has not been announced yet, or published (on Maven Central).  If the .war was uploaded to the Google storage bucket it can easily be deleted.

So, we should simply omit the links from the release notes and have a caveat somewhere that 2.16.11 was not released and to use 2.16.11.1 instead (if that is the number that we eventually use)

Patch-set: 1
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>