Document the known issue with Gerrit 2.15/NoteDb
Gerrit administrator need to be warned about the pothole
found after the release of Gerrit 2.15.

The migration to 2.16 solves the problem, however, if any of
the changes had a huge payload on a NoteDb review comment, the whole
migration could be impossible without any data loss.

It is best to document the known issue so that the administrators
can take an educated decision on when to migrate to NoteDb.

Bug: Issue 10239
Change-Id: Id5a5924717065431f40ff723262deb329c0f97e4