Update patch set 2
Patch Set 2:

(3 comments)

Thank you for the review.

Just for clarity:
I'm not pretending that I have the final interpretation of the relationship between repo-discuss/issue-tracker/design-feature-requests.
I uploaded this so that we could (hopefully) come to a mutual agreement of what we think is a good way to guide users with questions/issues/feature-requests.

I think that if we steer users to f.i. repo-discuss when that is more appropriate we cut down the flow to the issue-tracker and users get, on average, faster and better(!?) advice.

Patch-set: 2
Reviewer: Gerrit User 1014252 <1014252@173816e5-2b9a-37c3-8a2e-48639d4f1153>