Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> I think Luca's suggestions are nice, but wouldn't it be better to fix these in the source of this doc?  I think it's best if we take an existing standard CoC and follow their changes.

We started from the Google CoC for open source projects, but some parts of it are not applicable for Gerrit (e.g. it applies only if you officially represent the project and not for every contributor, violations can be escalated to some Google committee) which is why we couldn't take it as is. So we took most of it with some minor adoptions.

I've gone through many existing CoC's and none of them is perfect for us. Other projects like Chromium also decided to adapt their own variant of a CoC. So I think it's fine for us to do the same.

Patch-set: 9
Reviewer: Gerrit User 1003935 <1003935@173816e5-2b9a-37c3-8a2e-48639d4f1153>