Update patch set 2
Patch Set 2:

Note that this is still a draft; if there are additional items or further details on the items added here, they can still be added in follow up commits.  The post will not be published until after it's moved into the _posts folder and the date is 27th Sept or later.

Patch-set: 2
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>