Update patch set 11
Patch Set 11:

> Can we submit this?  It has +1 from 7 of the maintainers.

I would prefer to keep it open for a few more days so that everyone has a fair chance to see it and comment on it. We keep changes for design docs open for at least 10 days, so I think it makes sense to keep this change open for 10 days as well.

Patch-set: 11
Reviewer: Gerrit User 1003935 <1003935@173816e5-2b9a-37c3-8a2e-48639d4f1153>