Restructure ESC minutes
Rather than having each item in the minutes as a bullet item, use
headings. This makes it easier to skim the page and see an overview
of the topics discussed. It also results in anchors being generated
in the HTML allowing direct links to items (i.e. in emails, instant
messenger, etc). In future we could also generate tables of contents
on the posts.

Change-Id: I7cbeb705440e0be3b8ad0cea201749c2d3f8dded