Testimonials: Rename to stories as welcoming term
Rename 'testimonials' to 'user stories' as a Community resource, so it
sounds less formal. Do so in order to foster such contributions some
more from community members. Testimonials may mean endorsement or
recommendations for organizations, likely with impeding legal
implications internally. Agile did coin the term User story previously
of course, however it is still deemed as an improving one herein.

Clarify that stories (or usage feedback about Gerrit) shall remain
positive in tone, as this is the goal of that sharing section.

Slightly refactor the page, accordingly.

Change-Id: I548bb5d8645a8ff670927f1fef66645a7627c42b