Remove unused documentation files
Gerrit's documentation was added to the project during initial
experimentation with the new Jekyll templating system, but is
not actually published anywhere and has not been kept up to date
with the current documentation that is maintained in the core
gerrit project. It's better to remove it than leave it getting
even more stale.

Also remove the gerritdoc_sidebar definition. None of the pages
on the site are using it (they all explicitly set the hide_sidebar
option to true).

Change-Id: Id3ffac99e785750f7fae8cf6f98821dd36299e05