Update patch set 2
Patch Set 2:

(1 comment)

Thanks Edwin for the quick reply. I did not even realised that there was a bug open on that :-)
I just looked this morning *once again* the old GWT screenshot and was so sick of seeing it on the home page that I fixed it :-)

Adding David Ostrovsky as reviewer for his GDPR consent to have his name in the home page of Gerrit.

Patch-set: 2
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>