Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Why not find the same change that was used for the original screen shot and make a new one of the same review?

The old screenshot has also names of people that aren't working on the project anymore.
It would be even best to edit the page and avoid putting personal information altogether, because you never know what would happen in the future and who will join/leave the project.

Patch-set: 2