Update patch set 2
Patch Set 2: Code-Review-1

I do support the relief from Gerrit master of the expensive computation of the merge-ability flag, especially because it is just not scalable as the number of open changes on the same repo/branch grows.

People and Teams are moving to mono-repos trunk-based development and that would make Gerrit scalability the most critical point to address.

I am only concerned about the fact that people use that flag and the "is:mergeable" query very often: they do rely on it for avoiding the triggering of the builds, they do use it for avoiding reviewing a change that isn't mergeable anyway and thus would require rework.

I do see a one-way solution to the problem. However, in our design template, I believe we ask to provide a list of options that you have evaluated with the pros and cons.

Do you foresee other ways that Gerrit (or anything else) can provide the same functionality and keep on serving the 'is:mergeable' queries and showing the 'merge conflict' status?

Patch-set: 2
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1