Update patch set 5
Patch Set 5:

I reworked the use cases doc a bit and tried to answer your questions in a way that does not entail any solution.

I guess I look entirely foolish with this doc. Last time, I got a comment that the doc seemed prefabricated and that it appeared that there was no desire to make fundamental changes. This is why this time I wanted to start from scratch in the open, even if that comes at the expense of making me look foolish or making the effort look not well thought out.

It seems like a good idea to get the use cases agreed on and signed off first before I rework any solutions.

Patch-set: 5