Update patch set 1
Patch Set 1:

Thanks for your work on this

The idea is that this will better set expectations about bug handling: bugs that are in the roadmap or tied to an upcoming release will be P2 or higher (in particular, if a regression is discovered it would be in this category) but sometimes there are longstanding issues in a "help wanted" state and we want to be honest about that.

Also happy to see wording improvements over time too. :)

Patch-set: 1