Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Thanks for your work on this

Thanks for your reviews, I've learned something new :-)
I never heard of SLO and now I do know what it is.

> The idea is that this will better set expectations about bug handling: bugs that are in the roadmap or tied to an upcoming release will be P2 or higher (in particular, if a regression is discovered it would be in this category) but sometimes there are longstanding issues in a "help wanted" state and we want to be honest about that.

Yes, makes sense.

> Also happy to see wording improvements over time too. :)

Sure, this is just the starting point. At lease we set some initial expectations and clarify that it is not "us" and "them" but only "all of us" working and solving the issues.

Patch-set: 1