Update patch set 10
Patch Set 10:

(4 comments)

Apologies, I believe I misunderstood in my earlier post.
Now it is quite clear to me that the solution (horizontal scaling) should be part of the follow-up change and not of this one.

The use-case is clear: scaling up the replication. The solution, depending on the specs of the repos and bottlenecks could be different.

Patch-set: 10
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>