Update patch set 10
Patch Set 10:

> Patch Set 10:
> Now it is quite clear to me that the solution (horizontal scaling) should be part of the follow-up change and not of this one.
>
> The use-case is clear: scaling up the replication. The solution, depending on the specs of the repos and bottlenecks could be different.

I can see why you would think this. I believe that the use case as I have framed it is appropriate though, and I will ask you to take a step back and try to see how the way I am framing this is a pattern that we want to encourage in the future in some cases.

While you see adding more nodes as a solution, I am claiming that in this case it is NOT a solution but rather it is a way of using the system. I want adding more nodes to be a way to use the system effectively. Yes, using a tool a certain way may be a solution to another problem, however the solution to other problems is often abstracted to be a need in itself.

For example, we want Gerrit to be able to support large heap sizes, we accept that supporting large heaps sizes, while perhaps a "secondary need", is a valuable need in itself. In other words, we may be curious as to why users would need to use more heap, maybe in order to support more projects, more users, more operations... but we can understand that supporting a larger heap size will also allow Gerrit to support many future scaling problems that we have not yet even encountered. Thus supporting a larger heap size is something that we see as a need in itself and I will assert that we would want to support a use case that read something like this: "I want to add more heap to my Gerrit instance, however Gerrit crashes when I raise it above 10GB. The heap limit in Gerrit should be raised."

Today Gerrit can scale most operations horizontally by adding more nodes, for example we can scale downloads by adding slaves. Those with multi-master setups can handle more uploads and more WUI traffic by adding more nodes. The problem we are facing is that currently replication does not scale this way. The specifics to what bottlenecks someone is running into do not really matter in this particular case because the objective is to build a system that is future proof, i.e. to add support for a mode of operation that can handle a hole class of bottlenecks, real bottlenecks that exist today like the ones I have already described below, and future bottlenecks that don't even exist today. In other words, this use case is about making it possible for admins to scale replication when new bottlenecks arise even if the Gerrit project becomes defunct and is no longer maintained.

Patch-set: 10