Update patch set 6
Patch Set 6:

Following some discussion on the hackathon, we are proceeding with fixing the reindexAfterRefUpdate config by replacing it with indexMergeable. The change to do this is  I053af1b99616. That commit message contains a detailed description on why replacing reindexAfterRefUpdate with indexMergeable is more consistent and fixing behavior that is currently broken in Gerrit.

I'll abandon this doc soon. It served as a good discussion forum, but the change seems small enough that it doesn't require a design doc.

Patch-set: 6
Reviewer: Gerrit User 1026112 <1026112@173816e5-2b9a-37c3-8a2e-48639d4f1153>