Update patch set 5
Patch Set 5: Code-Review-1

I think it's a great feature to control which changes needs your attention in general. Setting up rules for which changes ends up in your to-do list is a long-standing feature request from a plethora of our users.

However I don't see why the feature requests makes it as if there's an either-or relationship between attention-sets and assignee.
AFAICT attention-sets aren't solving the problem that assignee attempts to solve of: "shared responsibility is no responsibility". Being added as reviewer or in an attention-set can (and are) easily ignored by users who feel that they don't have time to review ATM and that it doesn't matter because "some other reviewer" will.
If I need one specific person to review the change I cannot (AFAICT) express that with attention-sets. With the default behavior I would upload a change, remove all-but-one from the attention set. Even then there seem to be no way of that person to know that it's currently the only one in the attention-set from looking at the dashboard. And if/when I upload a new ps I need to once again go through the procedure of removing all-but-one of the reviewers and cc from the attention-set.
It wouldn't be that hard to implement some sort of default behavior like "if the assignee has replied remove the user as assignee". Implementing this and other behavior in a plugin is no problem either.

I would prefer if the removal of assignee is not part of this feature request or that the attention-set feature has some way of expressing the same single-user-responsibility that the assignee feature has.

Patch-set: 5
Reviewer: Gerrit User 1014252 <1014252@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1