Update patch set 5
Patch Set 5:

yes, i commented about diffusion of responsibility just a few hours ago. ^^

i have a more critical view of the configurability, but that was so far based on the fact that the design suggests making this an admin thing. if this is moved into the realm of users (with admin-provided defaults) it becomes theoretically less problematic. however, then there is a non-trivial chance that users get it wrong and start relying on those broken rules, which isn't better, either.

Patch-set: 5