Update patch set 5
Patch Set 5:

Hi Sven, thanks a lot for the feedback. I am eager to have this discussion and hopefully we can end up on the same page.

> I don't see why the feature requests makes it as if there's an either-or relationship between attention-sets and assignee.

I think it would be very confusing, if we had two competing systems that are trying to solve the "Whose turn is it?" use case.

Note that maybe we should drop the term "attention set" and just keep calling it "assignee". Then we are basically proposing to allow multiple users to be assigned at the same time.

> AFAICT attention-sets aren't solving the problem that assignee attempts to solve of: "shared responsibility is no responsibility". Being added as reviewer or in an attention-set can (and are) easily ignored by users who feel that they don't have time to review ATM and that it doesn't matter because "some other reviewer" will.

If an owner adds more than one reviewer, but only expects action from on of them, then why are they adding more reviewers in the first place?

I agree that without attention set and without assignee there can be the attitude of "I don't feel responsible and hope someone else reviews the change." We want to eliminate this attitude entirely. If you are assigned, then it is your turn. The owner knows it, everyone else knows it. The most prominent section of the dashboard tells you so. The change page clearly indicates it. You cannot duck away from it by being silent. The attention set assignment clearly points to you. If you don't have time for the review, then you can remove yourself from the attention set. So we are turning around the responsibility: The owner does not have to chase or ping reviewers, but the reviewers have to act, if they are not planning to review something.

> If I need one specific person to review the change I cannot (AFAICT) express that with attention-sets.

You can. Just assign the review to this single person.

> With the default behavior I would upload a change, remove all-but-one from the attention set.

Why do you add them as reviewers in the first place, if you don't want a review from them?

> Even then there seem to be no way of that person to know that it's currently the only one in the attention-set from looking at the dashboard.

There is! It should be very clear to everyone on the dashboard and on the change page, who is currently assigned and who is not.

> And if/when I upload a new ps I need to once again go through the procedure of removing all-but-one of the reviewers and cc from the attention-set.

Again, why are others reviewers, if you don't want anything from them?

CCs are not automatically added to the attention set. I have either poorly written the rule or you have misunderstood it.

> It wouldn't be that hard to implement some sort of default behavior like "if the assignee has replied remove the user as assignee".

Yes, you can treat my proposal of just an evolvement of the assignee feature. It is actually not much more than "allow multiple assignees", "ease the back and forth with automation" and "more UI emphasis".

> Implementing this and other behavior in a plugin is no problem either.

Technically correct, apart from maybe the headache of indexing plugin related data.

The important point is that knowing whose turn it is in a turn based workflow going back and forth between owner and reviewers is not a plugin feature, but an absolutely fundamental part of the code review product. If we want to establish this new concept for all users, then we need to fully back it up and make it part of core, and e.g. add these ideas to the front page of our code review documentation.
 
> I would prefer if the removal of assignee is not part of this feature request or that the attention-set feature has some way of expressing the same single-user-responsibility that the assignee feature has.

Again, if you want to single out one person, then there is nothing stopping you from just putting one person into the attention set.

I think the important missing part maybe is that we expect all companies using Gerrit to send an announcement when the attention set feature lands: "Listen everybody, starting today you are expected to act on every change with you in the attention set!"

Patch-set: 5