Update patch set 5
Patch Set 5:

I'm 95% positive to the feature request and I'm quite confident that it will result in an improvement.
My main concern is that I see great value in identifying one primary responsible and I don't see how AS supports this use-case at it currently stands.

> I think it would be very confusing, if we had two competing systems that are trying to solve the "Whose turn is it?" use case.
Agreed. I think we should let AS support singling out one reviewer.

> > If I need one specific person to review the change I cannot (AFAICT) express that with attention-sets.
> 
> You can. Just assign the review to this single person.

But if someone else adds a second person to the attention-set it is back to being a shared responsibility until you remove the added user.

> > Even then there seem to be no way of that person to know that it's currently the only one in the attention-set from looking at the dashboard.
> 
> There is! It should be very clear to everyone on the dashboard and on the change page, who is currently assigned and who is not.

Are all users in the attention-set visible in the dashboards? I fear that this can get ugly pretty quick.
I also feel that who else is in the AS shouldn't concern reviewers to any large degree. It opens up for someone-else-ism, "Oh, there's three people in the attention-set of that change. Good, then I don't even have to look at it".

> > Implementing this and other behavior in a plugin is no problem either.
> 
> Technically correct, apart from maybe the headache of indexing plugin related data.

Just to be perfectly clear; I did *not* mean to implement the Attention Set feature in a plugin. That would be craziness.
What I meant was to implement the behavior of assignee or AS in a plugin. "If someone has left a change-comment remove them from the AS/assignee", that type of thing.

> I think the important missing part maybe is that we expect all companies using Gerrit to send an announcement when the attention set feature lands: "Listen everybody, starting today you are expected to act on every change with you in the attention set!"

If that could be accomplished effectively through an email you could do that today. Just switch "attention set" for "reviewer".



Hopefully this somewhat clears up my view on the other questions raised:

1. Having the assignee as a set of reviewers would totally defeat it's purpose.
The only purpose of assignee is that it singles out one single reviewer and says to that reviewer "You are expected to review this change".
The Attention Set's "one of you three reviewers are expected to review this change" opens up for an implicit negotiation of who that "one of you" should be and the answer is surprisingly often "someone else".

2. If I squint and look upon this feature request unfavorably I could say that it basically adds an abstraction around reviewers but it's essentially the same thing. If the solution is to add only one reviewer so that this sole reviewer adds up in the AS, why do we need an AS? Then the current reviewer field is good enough? Adding only one reviewer without the AS would have the same effect. I personally see the direct correlation between reviewer<->Attention Set member as one of the weak-points in this feature request.
If they aren't the same thing and the users in the attention-set are the only ones who are expected to review the change, what does it mean to be added as a reviewer?

3. It's quite common to add more reviewers than you need to review the change.
You might have a pool of people that could review the change but you have no clear picture of who's currently available. If you only add one of those reviewers you have to wait until that reviewer reacts or until a timeout is reached where you remove that reviewer and add another one from the pool and repeat the process.
Adding a set of reviewers gives them a heads-up so that they can plan their reviews when they go into review-mode.
We have a two-pair-of-eyes policy which implicitly make CR+1 AND CR+2 a submit criteria. You might have one individual that you need to get a review from while the other reviewer can be whoever from a set.
We have come to rely quite heavily on the reviewers plugin. It helps us a lot because a common use-case is to make some changes in an area of the platform that is someone else's responsibility and you might not have a close enough relationship with that team to know who you should add as a reviewer. And there might not be an easy way for that team to automatically single out one, available, reviewer.

I will revisit the feature request when I'm not as incapacitated.
I have quite possibly missed something in the feature request but have you considered:

* Removing CC, and having only two sets, reviewers and attention-set?
Where being added as a Reviewer means: "This might concern you", and being added to the attention-set means: "You are expected to review this".
In the UI you could move user between the two sets through drag-and-drop.

* Adding the possibility to label one of the users in the attention-set as "primary" in some way. This might be enough to serve the same purpose as the assignee field does today.

Patch-set: 5
Reviewer: Gerrit User 1014252 <1014252@173816e5-2b9a-37c3-8a2e-48639d4f1153>